tootalooBackend/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
Django==4.1.6
django-environ==0.9.0
dnspython==2.3.0
mongoengine==0.26.0
pymongo==4.3.3
sqlparse==0.4.3
